eng/misc/fn-misc-ipchecker/src/requirements.txt
antlr4-python3-runtime==4.9.3 ; python_version >= "3.10" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.4.0 ; python_version >= "3.10" and python_version < "4.0"
google-api-core==2.22.0 ; python_version >= "3.10" and python_version < "4.0"
google-api-core[grpc]==2.22.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.35.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-bigquery==3.26.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.10" and python_version < "4.0"
google-crc32c==1.6.0 ; python_version >= "3.10" and python_version < "4.0"
google-resumable-media==2.7.2 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.65.0 ; python_version >= "3.10" and python_version < "4.0"
grpcio-status==1.67.1 ; python_version < "4.0" and python_version >= "3.10"
grpcio==1.67.1 ; python_version < "4.0" and python_version >= "3.10"
hydra-core==1.3.2 ; python_version >= "3.10" and python_version < "4.0"
idna==3.10 ; python_version >= "3.10" and python_version < "4.0"
numpy==2.1.3 ; python_version >= "3.10" and python_version < "4.0"
omegaconf==2.3.0 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.10" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.10" and python_version < "4.0"
proto-plus==1.25.0 ; python_version >= "3.10" and python_version < "4.0"
protobuf==5.28.3 ; python_version >= "3.10" and python_version < "4.0"
pyarrow==17.0.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
pytz==2024.2 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2024.2 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.10" and python_version < "4.0"
